[collection] refactoring, moving parts out of main.rs to module
