+++ src/main.rs
@@ -1,30 +1,31 @@
 //! A simple first program in rust, actually I plan to do something even
 //! useful. Trying to find all my audiobooks on many machines, identify them,
-//! find duplicates (later trying to solve the problem, also including their permissions, different names, but same albums, etc),
-//! get all stats about it).
+//! find duplicates (later trying to solve the problem, also including their permissions,
+//! different names, but same albums, etc, get all stats about it).
 //! It acts as a wrapper around adbflib, which holds major parts of the implemention.
 extern crate adbflib;
 extern crate rayon;
 
 use adbflib::{
     command_line,
-    ctrl::{CollectionPathAlive, Ctrl, ForwardNetMessage, NetMessages, Status, UiUpdateMsg},
-    data::{self, ipc::IPC, Collection, Container},
+    ctrl::{Ctrl, UiUpdateMsg},
+    data::{
+        self,
+        collection::{Collection, Container},
+        ipc::IPC,
+        CollectionOutputData,
+    },
     logit,
     net::{key_keeper, Net},
 };
 
 use async_std::task;
 use crossbeam::{channel::unbounded, sync::WaitGroup};
-use log::{error, info, trace, warn};
+use log::{error, info, trace};
 use rayon::prelude::{IndexedParallelIterator, IntoParallelRefIterator, ParallelIterator};
 use std::{
     io::{self, Error},
-    path::Path, // path, clear
-    sync::{
-        mpsc::{channel, Sender},
-        Arc as SArc, Mutex as SMutex,
-    },
+    sync::{mpsc::channel, Arc as SArc, Mutex as SMutex},
 };
 
 fn main() -> io::Result<()> {
@@ -128,13 +129,13 @@ fn main() -> io::Result<()> {
         let init_collection = Collection::new(&key_keeper::get_p2p_server_id(), max_threads);
         let collection_protected = SArc::new(SMutex::new(init_collection));
 
-        let output_data = SArc::new(SMutex::new(CollectionData { nr_found_songs: 0 }));
+        let output_data = SArc::new(SMutex::new(CollectionOutputData { nr_found_songs: 0 }));
         let handle_container = SArc::new(SMutex::new(Container::new()));
 
         &ui_paths.par_iter().enumerate().for_each(|(index, elem)| {
             let sender_loop = synced_to_ui_messages.clone();
             let collection_data_in_iterator = collection_protected.clone();
-            let single_path_collection_data = search_in_single_path(
+            let single_path_collection_data = data::search_in_single_path(
                 has_ui,
                 handle_container.clone(),
                 collection_data_in_iterator,
@@ -210,106 +211,4 @@ fn main() -> io::Result<()> {
         drop(ui_thread);
     }
     Ok(())
-}
-
-// todo: move all below into data module
-struct CollectionData {
-    nr_found_songs: u32,
-}
-
-fn search_in_single_path(
-    has_ui: bool,
-    collection_data: SArc<SMutex<Container>>,
-    collection_protected: SArc<SMutex<Collection>>,
-    mutex_to_ui_msg: SArc<SMutex<Sender<UiUpdateMsg>>>,
-    index: usize,
-    elem: &str,
-) -> CollectionData {
-    if !has_ui {
-        println!("[{:?}] looking into path {:?}", index, elem);
-    } else {
-        // send start animation for that path
-        mutex_to_ui_msg
-            .lock()
-            .and_then(|locked_to_start_ui_message| {
-                trace!("send startAnimation for path {:?}", index);
-                locked_to_start_ui_message
-                    .send(UiUpdateMsg::CollectionUpdate(
-                        CollectionPathAlive::BusyPath(index),
-                        Status::ON,
-                    ))
-                    .unwrap_or_else(|_| warn!("... lost start animation for index {:?}", index));
-                trace!("start busy animation");
-                Ok(())
-            })
-            .unwrap_or_else(|_| error!("... that should not happen here at start"));
-    }
-    // todo: crap : unlock here is stupid!
-    let locked_collection = &mut *collection_protected.lock().unwrap();
-
-    // do it: main task here is to visit and dive deep
-    //        into the subfolders of this folder
-    match locked_collection.visit_path(
-        collection_data,
-        Path::new(elem),
-        &data::Collection::visit_files,
-    ) {
-        Ok(local_stats) => {
-            if has_ui {
-                // send stop animation for that path
-                trace!("send stopAnimation for path {:?}", index);
-                mutex_to_ui_msg
-                    .lock()
-                    .and_then(|locked_to_stop_ui_message| {
-                        locked_to_stop_ui_message
-                            .send(UiUpdateMsg::CollectionUpdate(
-                                CollectionPathAlive::BusyPath(index),
-                                Status::OFF,
-                            ))
-                            .unwrap_or_else(|_| warn!("... lost stop animation for {:?}", index));
-                        trace!("stop busy animation");
-                        Ok(())
-                    })
-                    .unwrap_or_else(|_| {
-                        error!("... that should not happen here at stop");
-                    });
-            } else {
-                let text = format!(
-                    "\n\
-                             analyzed: {an:>width$}, faulty: {fa:>width$}\n\
-                             searched: {se:>width$}, other: {ot:>width$}",
-                    an = local_stats.analyzed,
-                    fa = local_stats.faulty,
-                    se = local_stats.searched,
-                    ot = local_stats.other,
-                    width = 3
-                );
-                println!("[{:?}] done {}", index, text);
-            }
-            // return this here
-            CollectionData {
-                nr_found_songs: local_stats.analyzed,
-            }
-        }
-        Err(_e) => {
-            let text = format!("An error has occurred in search path [{}]!!", index);
-            if has_ui {
-                mutex_to_ui_msg
-                    .lock()
-                    .and_then(|locked_update_message| {
-                        locked_update_message
-                            .send(UiUpdateMsg::NetUpdate(ForwardNetMessage::Stats(
-                                NetMessages::Debug(text),
-                            )))
-                            .unwrap();
-                        Ok(())
-                    })
-                    .unwrap();
-            } else {
-                println!("{:?}", text);
-            }
-            // return this here
-            CollectionData { nr_found_songs: 0 }
-        }
-    }
 }
